Gameplay: extend zones, boss, loot & run task
Bump version to v1.19 and expand world/encounter logic: zones and boss thresholds moved (final boss now at Y>=240), enemy pools enriched with varied spawns and elite chances, and final-boss phases adjusted. Improve loot and relic handling (zone ranges, safer random indexing, weapon stat rounding) and potion/heal math. Add respawn/auto-heal behavior, status UI tweaks, initial story/intro and chest choice, and ensure player starts with appropriate starter gear. Editor/IDE changes: switch debugger console to externalTerminal, add a "run" task and Run.bat to launch the game. Misc: small API/typing fixes (static_cast usage, lowercase helper) and other minor safety/clarity improvements.

+++ Tienda.cpp
@@ -9,6 +9,8 @@ using namespace std;
 // BLOQUE M3: SISTEMA DE TIENDA (v1.18)
 // =========================================================
 
+
+//Funcion para entrar a la tienda
 void entrarTienda(Personaje &p) {
     bool enMenu = true;
 
@@ -83,8 +85,8 @@ void entrarTienda(Personaje &p) {
             cout << idx << ". Cancelar" << endl;
             int venta; cin >> venta;
             if (venta > 0 && venta < idx) {
-                string elegido = opcionesVenta[venta-1].first;
-                int precio = opcionesVenta[venta-1].second;
+                string elegido = opcionesVenta[static_cast<size_t>(venta-1)].first;
+                int precio = opcionesVenta[static_cast<size_t>(venta-1)].second;
 
                 auto itInv = find(p.inventario.begin(), p.inventario.end(), elegido);
                 if (itInv != p.inventario.end()) p.inventario.erase(itInv);
